test: Refactor test set up for readability (#810)
