+++ test/src/tests-forwarders.js
@@ -1,9 +1,8 @@
-import Utils from './utils';
+import Utils from './config/utils';
 import sinon from 'sinon';
 import fetchMock from 'fetch-mock/esm/client';
 
-import { urls } from './config';
-import { apiKey, MPConfig, testMPID, MessageType } from './config';
+import { urls, apiKey, MPConfig, testMPID, MessageType } from './config/constants';
 import { expect } from 'chai';
 
 const findEventFromRequest = Utils.findEventFromRequest,
