+++ test/src/tests-identities-attributes.js
@@ -1,9 +1,9 @@
-import Utils from './utils';
+import Utils from './config/utils';
 import sinon from 'sinon';
 import fetchMock from 'fetch-mock/esm/client';
 import { urls, apiKey,
     testMPID,
-    MPConfig } from './config';
+    MPConfig } from './config/constants';
 
 const findEventFromRequest = Utils.findEventFromRequest,
     findBatch = Utils.findBatch,
