+++ test/src/tests-self-hosting-specific.js
@@ -1,7 +1,7 @@
-import Utils from './utils';
+import Utils from './config/utils';
 import sinon from 'sinon';
 import fetchMock from 'fetch-mock/esm/client';
-import { urls, apiKey, MPConfig } from './config';
+import { urls, apiKey, MPConfig } from './config/constants';
 
 const { findEventFromRequest, findBatch } = Utils;
