--- test/test-index.js
@@ -1,5 +1,0 @@
-window.MockHttpServer = require('./mockhttprequest.js');
-window.Should = require('should');
-window.sinon = require('../node_modules/sinon/lib/sinon.js');
-require('./geomock.js');
-require('./src/tests-main.js');
